% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{coef.lp}
\alias{coef.lp}
\alias{coef.mf}
\title{Coefficients}
\usage{
\method{coef}{lp}(object, ...)

\method{coef}{mf}(object, ...)
}
\arguments{
\item{object}{object of class 'transreg'}

\item{...}{(not applicable)}
}
\description{
Extracts coefficients
}
\examples{
NA

}
