% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{coef.transreg}
\alias{coef.transreg}
\title{Coefficients}
\usage{
\method{coef}{transreg}(object, stack = NULL, ...)
}
\arguments{
\item{object}{object of class \code{transreg}}

\item{stack}{character "lp" (linear predictor stacking) or "mf" (meta-feature stacking)}

\item{...}{(not applicable)}
}
\description{
Extracts coefficients
}
\examples{
NA
}
