% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{cv.transfer}
\alias{cv.transfer}
\title{Cross-validation}
\usage{
cv.transfer(
  target,
  source = NULL,
  prior = NULL,
  z = NULL,
  family,
  alpha,
  scale,
  sign,
  select,
  switch,
  foldid.ext = NULL,
  nfolds.ext = 10,
  foldid.int = NULL,
  nfolds.int = 10,
  type.measure = "deviance",
  alpha.prior = NULL,
  partitions = NULL,
  monotone = NULL
)
}
\arguments{
\item{target}{list with slot x (feature matrix with n rows and p columns) and slot y (target vector of length n)}

\item{source}{list of k lists, each with slot x (feature matrix with m_i rows and p columns) and slot y (target vector of length m_i)}

\item{prior}{prior coefficients: matrix with \eqn{p} rows and \eqn{k} columns}

\item{z}{prior weights}

\item{family}{character "gaussian", "binomial", or "poisson"}

\item{alpha}{elastic net mixing parameter (0=ridge, 1=lasso)}

\item{scale}{"exp" for exponential scaling (3 parameters),
"iso" for isotonic scaling (1+p parameters)}

\item{sign}{sign discovery procedure: logical}

\item{select}{select from sources: logical}

\item{switch}{choose between positive and negative weights for each source: logical}

\item{foldid.ext}{external fold identifiers}

\item{nfolds.ext}{number of external folds}

\item{foldid.int}{internal fold identifiers}

\item{nfolds.int}{number of internal folds}

\item{type.measure}{character}

\item{alpha.prior}{alpha for source regression}

\item{partitions}{monotone: for GRridge}

\item{monotone}{logical}
}
\description{
Performs external \eqn{k}-fold cross-validation
to estimate the predictive performance of different methods
}
\examples{
NA

}
