% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{.calibrate}
\alias{.calibrate}
\title{Internal functions}
\arguments{
\item{plot}{logical}
}
\description{
Internal functions called by
\code{\link[=transreg]{transreg()}},
depending on choice between
exponential and isotonic calibration.
}
\seealso{
Use \code{\link[=transreg]{transreg()}} for model fitting.
}
