% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{.extract}
\alias{.extract}
\title{Internal functions}
\description{
Internal functions called by
\code{\link[=coef.transreg]{coef.transreg()}}, \code{\link[=predict.transreg]{predict.transreg()}} and \code{\link[=weights.transreg]{weights.transreg()}},
depending on choice between
linear predictor stacking
and meta-feature stacking.
}
\seealso{
Use \code{\link[=coef.transreg]{coef}},
\code{\link[=predict.transreg]{predict}}
and \code{\link[=weights.transreg]{weights}}.
}
