% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{.iso.slow.single}
\alias{.iso.slow.single}
\title{Isotonic scaling}
\usage{
.iso.slow.single(y, X, prior, family)
}
\arguments{
\item{y}{target: vector of length \eqn{n} (see \code{family})}

\item{X}{features: matrix with \eqn{n} rows (samples)
and \eqn{p} columns (features)}

\item{prior}{prior coefficients: matrix with \eqn{p} rows (features)
and \eqn{k} columns (sources of co-data)}

\item{family}{character "gaussian" (\eqn{y}: real numbers),
"binomial" (\eqn{y}: 0s and 1s),
or "poisson" (\eqn{y}: non-negative integers);}
}
\description{
Performs isotonic scaling. This function is for comparison only.
}
\examples{
NA

}
\seealso{
\code{\link[=.iso.fast.single]{.iso.fast.single()}}
}
