% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{.predict.lp}
\alias{.predict.lp}
\title{Internal functions}
\usage{
.predict.lp(object, newx, ...)
}
\arguments{
\item{object}{object of class \code{transreg}}

\item{newx}{features:
matrix with \eqn{n} rows (samples) and \eqn{p} columns (variables)}

\item{...}{(not applicable)}
}
\description{
Internal functions called by
\code{\link[=coef.transreg]{coef.transreg()}} and \code{\link[=predict.transreg]{predict.transreg()}},
depending on choice between
linear predictor stacking
and meta-feature stacking.
}
\examples{
NA
}
\seealso{
Use \code{\link[=coef.transreg]{coef}}
and \code{\link[=predict.transreg]{predict}}.
}
