% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{.signdisc}
\alias{.signdisc}
\title{Sign discovery}
\usage{
.signdisc(y, X, prior, family, foldid = NULL, nfolds = 10, track = FALSE)
}
\arguments{
\item{y}{target: vector of length \eqn{n} (see \code{family})}

\item{X}{features: matrix with \eqn{n} rows (samples)
and \eqn{p} columns (features)}

\item{prior}{prior coefficients: matrix with \eqn{p} rows (features)
and \eqn{k} columns (sources of co-data)}

\item{family}{character "gaussian" (\eqn{y}: real numbers),
"binomial" (\eqn{y}: 0s and 1s),
or "poisson" (\eqn{y}: non-negative integers);}

\item{foldid}{fold identifiers: vector of length \eqn{n}
with entries from 1 to `nfolds`}

\item{nfolds}{number of folds: positive integer}

\item{track}{show intermediate output (messages and plots): logical}
}
\description{
Assigns signs to prior weights to obtain prior coefficients
}
