% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{exp.multiple}
\alias{exp.multiple}
\title{Exponential scaling}
\usage{
\method{exp}{multiple}(y, X, prior, family, select, plot = TRUE)
}
\arguments{
\item{y}{target: vector of length \eqn{n}}

\item{X}{features: matrix with \eqn{n} rows and \eqn{p} columns}

\item{prior}{prior coefficients: matrix with \eqn{p} rows and \eqn{k} columns}

\item{family}{character "gaussian", "binomial", or "poisson"}

\item{select}{select from sources: logical}

\item{plot}{logical}
}
\description{
Performs exponential scaling
}
\examples{
NA

}
