% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{.which.stack}
\alias{.which.stack}
\alias{extract}
\alias{.predict.lp}
\alias{.predict.mf}
\alias{.coef.lp}
\alias{.coef.mf}
\alias{.weights.lp}
\alias{.weights.mf}
\title{Internal functions}
\usage{
.which.stack(object, stack)

.predict.lp(object, newx, ...)

.predict.mf(object, newx, ...)

.coef.lp(object, ...)

.coef.mf(object, ...)

.weights.lp(object, ...)

.weights.mf(object, ...)
}
\arguments{
\item{object}{object of class \code{transreg}}

\item{stack}{character "lp" (linear predictor stacking) or "mf" (meta-feature stacking)}

\item{newx}{features:
matrix with \eqn{n} rows (samples) and \eqn{p} columns (variables)}

\item{...}{(not applicable)}
}
\description{
Internal functions called by
\code{\link[=coef.transreg]{coef.transreg()}}, \code{\link[=predict.transreg]{predict.transreg()}} and \code{\link[=weights.transreg]{weights.transreg()}},
depending on choice between
linear predictor stacking
and meta-feature stacking.
}
\section{Functions}{
\itemize{
\item \code{.which.stack}: called by \code{coef.transreg}, \code{predict.transreg} and \code{weights.transreg}

\item \code{.predict.lp}: called by \code{predict.transreg} if \code{stack="lp"}

\item \code{.predict.mf}: called by \code{predict.transreg} if \code{stack="mf"}

\item \code{.coef.lp}: called by \code{coef.transreg} if \code{stack="lp"}

\item \code{.coef.mf}: called by \code{coef.transreg} if \code{stack="mf"}

\item \code{.weights.lp}: called by \code{weights.transreg} if \code{stack="lp"}

\item \code{.weights.mf}: called by \code{weights.transreg} if \code{stack="mf"}
}}

\seealso{
Use \code{\link[=coef.transreg]{coef}},
\code{\link[=predict.transreg]{predict}}
and \code{\link[=weights.transreg]{weights}}.
}
