% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{iso.multiple}
\alias{iso.multiple}
\title{Isotonic scaling}
\usage{
iso.multiple(y, X, prior, family, select = TRUE, switch = TRUE)
}
\arguments{
\item{y}{target: vector of length \eqn{n}}

\item{X}{features: matrix with \eqn{n} rows and \eqn{p} columns}

\item{prior}{prior coefficients: matrix with \eqn{p} rows and \eqn{k} columns}

\item{family}{character "gaussian", "binomial", or "poisson"}

\item{select}{select from sources: logical}

\item{switch}{choose between positive and negative weights for each source: logical}
}
\description{
Performs isotonic scaling
}
\examples{
NA

}
