% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{iso.slow.single}
\alias{iso.slow.single}
\title{Isotonic scaling}
\usage{
iso.slow.single(y, X, prior, family)
}
\arguments{
\item{y}{target: vector of length \eqn{n}}

\item{X}{features: matrix with \eqn{n} rows and \eqn{p} columns}

\item{prior}{prior coefficients: matrix with \eqn{p} rows and \eqn{k} columns}

\item{family}{character "gaussian", "binomial", or "poisson"}
}
\description{
Performs isotonic scaling. This function is for comparison only.
}
\examples{
NA

}
\seealso{
iso.fast.single
}
