% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{predict.lp}
\alias{predict.lp}
\alias{predict.mf}
\title{Predict}
\usage{
\method{predict}{lp}(object, newx, ...)

\method{predict}{mf}(object, newx, ...)
}
\arguments{
\item{object}{object of class 'transreg'}

\item{newx}{features: matrix with m rows (samples) and p columns (variables)}

\item{...}{(not applicable)}
}
\description{
Predicts outcome
}
\examples{
NA

}
