% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{predict.transreg}
\alias{predict.transreg}
\title{Predict}
\usage{
\method{predict}{transreg}(object, newx, stack = NULL, ...)
}
\arguments{
\item{object}{object of class \code{transreg}}

\item{newx}{features:
matrix with \eqn{n} rows (samples) and \eqn{p} columns (variables)}

\item{stack}{character "lp" (linear predictor stacking) or "mf" (meta-feature stacking)}

\item{...}{(not applicable)}
}
\description{
Predicts outcome
}
\examples{
NA
}
