% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{print.transreg}
\alias{print.transreg}
\title{Print transreg-object}
\usage{
\method{print}{transreg}(x, ...)
}
\arguments{
\item{x}{object of class transreg}

\item{...}{(not applicable)}
}
\description{
Show summary of transreg-object
}
\examples{
NA

}
