% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{residuals}
\alias{residuals}
\title{Calculate residuals}
\usage{
residuals(y, y_hat, family)
}
\arguments{
\item{y}{target: vector of length \eqn{n}}

\item{y_hat}{predicted values: vector of length n, or matrix with n rows and k columns}

\item{family}{"gaussian" or "binomial"}
}
\description{
Calculates residuals from observed outcome
and predicted values (Gaussian family)
or predicted probablities (binomial family).
}
\examples{
NA

}
