% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sign.disc}
\alias{sign.disc}
\title{Sign discovery}
\usage{
\method{sign}{disc}(y, X, prior, family, foldid = NULL, nfolds = 10)
}
\arguments{
\item{y}{target: vector of length \eqn{n}}

\item{X}{features: matrix with \eqn{n} rows and \eqn{p} columns}

\item{prior}{prior coefficients: matrix with \eqn{p} rows and \eqn{k} columns}

\item{family}{character "gaussian", "binomial", or "poisson"}

\item{foldid}{fold identifiers: vector of length \eqn{n} with entries from 1 to \code{nfolds}}

\item{nfolds}{number of folds: integer}
}
\description{
Assigns signs to prior weights to obtain prior coefficients
}
\examples{
NA

}
