% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{simulate}
\alias{simulate}
\title{Simulation}
\usage{
simulate(
  p = 1000,
  n.target = 100,
  n.source = 150,
  k = 3,
  family = "gaussian",
  prop = 0.01,
  rho.beta = 0.95,
  rho.x = 0.95,
  w = 0.5
)
}
\arguments{
\item{p}{number of features}

\item{n.target}{sample size for target data set}

\item{n.source}{sample size(s) for source data set(s), scalar or vector of length k}

\item{k}{number of source data sets}

\item{family}{"Gaussian", "binomial" or "poisson"}

\item{prop}{approximate proportion of features with effects}

\item{rho.beta}{correlation between effects (across different data sets)}

\item{rho.x}{base for decreasing correlation structure for correlation between features}

\item{w}{weight between signal and noise}
}
\description{
Simulates data
}
\examples{
NA

}
