% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{coef.transreg}
\alias{coef.transreg}
\title{Coefficients}
\usage{
\method{coef}{transreg}(object, ...)
}
\arguments{
\item{object}{object of class 'transreg'}

\item{...}{(not applicable)}
}
\description{
Extracts coefficients
}
\examples{
NA

}
