% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{extract}
\alias{extract}
\alias{.predict.lp}
\alias{.predict.mf}
\alias{.coef.lp}
\alias{.coef.mf}
\alias{.which.stack}
\title{Internal functions}
\usage{
.predict.lp(object, newx, ...)

.predict.mf(object, newx, ...)

.coef.lp(object, ...)

.coef.mf(object, ...)

.which.stack(object, stack)
}
\arguments{
\item{object}{object of class \code{transreg}}

\item{newx}{features:
matrix with \eqn{n} rows (samples) and \eqn{p} columns (variables)}

\item{...}{(not applicable)}

\item{stack}{character "lp" (linear predictor stacking) or "mf" (meta-feature stacking)}
}
\description{
Internal functions called by
\code{\link[=coef.transreg]{coef.transreg()}} and \code{\link[=predict.transreg]{predict.transreg()}},
depending on choice between
linear predictor stacking
and meta-feature stacking.
}
\section{Functions}{
\itemize{
\item \code{.predict.lp}: called by \code{predict.transreg} if \code{stack="lp"}

\item \code{.predict.mf}: called by \code{predict.transreg} if \code{stack="mf"}

\item \code{.coef.lp}: called by \code{coef.transreg} if \code{stack="lp"}

\item \code{.coef.mf}: called by \code{coef.transreg} if \code{stack="mf"}

\item \code{.which.stack}: called by \code{coef.transreg} and \code{predict.transreg}
}}

\seealso{
Use \code{\link[=coef.transreg]{coef}}
and \code{\link[=predict.transreg]{predict}}.
}
