% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{iso.multiple}
\alias{iso.multiple}
\title{Isotonic scaling}
\usage{
iso.multiple(y, X, prior, family, select = TRUE, switch = TRUE)
}
\arguments{
\item{y}{target: vector of length \eqn{n} (see \code{family})}

\item{X}{features: matrix with \eqn{n} rows (samples)
and \eqn{p} columns (features)}

\item{prior}{prior coefficients: matrix with \eqn{p} rows (features)
and \eqn{k} columns (sources of co-data)}

\item{family}{character "gaussian" (\eqn{y}: real numbers),
"binomial" (\eqn{y}: 0s and 1s),
or "poisson" (\eqn{y}: non-negative integers);}

\item{select}{select from sources: logical}

\item{switch}{choose between positive and negative weights for each source: logical}
}
\description{
Performs isotonic scaling
}
\examples{
NA

}
