% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{predict.transreg}
\alias{predict.transreg}
\title{Predict}
\usage{
\method{predict}{transreg}(object, newx, ...)
}
\arguments{
\item{object}{object of class 'transreg'}

\item{newx}{features: matrix with m rows (samples) and p columns (variables)}

\item{...}{not applicable}
}
\description{
Predicts outcome
}
\examples{
NA

}
