% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{residuals}
\alias{residuals}
\title{Calculate residuals}
\usage{
residuals(y, y_hat, family)
}
\arguments{
\item{y}{response: vector of length \eqn{n} (see family)}

\item{y_hat}{predicted values: vector of length \eqn{n},
or matrix with \eqn{n} rows and \eqn{k} columns}

\item{family}{character
"gaussian" (\eqn{y}: real numbers, \eqn{y_hat}: real numbers)
or "binomial" (\eqn{y}: 0s and 1s, \eqn{y_hat}: ) $0 \leq x \leq 1$}
}
\description{
Calculates residuals from observed outcome
and predicted values (Gaussian family)
or predicted probabilities (binomial family).
}
\examples{
NA

}
