% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sign.disc}
\alias{sign.disc}
\title{Sign discovery}
\usage{
\method{sign}{disc}(y, X, prior, family, foldid = NULL, nfolds = 10)
}
\arguments{
\item{y}{target: vector of length \eqn{n} (see \code{family})}

\item{X}{features: matrix with \eqn{n} rows (samples)
and \eqn{p} columns (features)}

\item{prior}{prior coefficients: matrix with \eqn{p} rows (features)
and \eqn{k} columns (sources of co-data)}

\item{family}{character "gaussian" (\eqn{y}: real numbers),
"binomial" (\eqn{y}: 0s and 1s),
or "poisson" (\eqn{y}: non-negative integers);}

\item{foldid}{fold identifiers: vector of length \eqn{n}
with entries from 1 to \code{nfolds}}

\item{nfolds}{number of folds: positive integer}
}
\description{
Assigns signs to prior weights to obtain prior coefficients
}
\examples{
n <- 100; p <- 500
X <- matrix(stats::rnorm(n*p),nrow=n,ncol=p)
beta <- stats::rnorm(p)*stats::rbinom(n=p,size=1,prob=0.2)
y <- X \%*\% beta
prior <- matrix(abs(beta),ncol=1)
temp <- sign.disc(y,X,prior,family="gaussian")
table(sign(beta),sign(temp))

}
