% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgname.R
\docType{package}
\name{transreg-package}
\alias{transreg-package}
\title{Penalised regression with multiple sets of prior effects}
\description{
The R package \code{starnet} implements penalised regression
with multiple sets of prior effects.
}
\details{
Use function \code{\link{transreg}} for model fitting.
Type \code{library(transreg)} and then \code{?transreg} or
\code{help("transreg)"} to open its help file.

See the vignette for further examples.
Type \code{vignette("transreg")} or \code{browseVignettes("transreg")}
to open the vignette.
}
\examples{
NA

}
\references{
Armin Rauschenberger, Zied Landoulsi,
Mark A van de Wiel, and Enrico Glaab (2022).
"Penalised regression with multiple sets of prior effects".
\emph{Manuscript in preparation.}
\email{armin.rauschenberger@uni.lu}
}
\keyword{documentation}
