% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgname.R
\docType{package}
\name{transreg-package}
\alias{transreg-package}
\title{Penalised regression with multiple sets of prior effects}
\description{
The R package `transreg` implements penalised regression
with multiple sets of prior effects.
}
\details{
Use function [transreg()] for model fitting.
Type `library(transreg)` and then `?transreg` or
`help("transreg")` to open its help file.

See the vignette for further examples.
Type `vignette("transreg")` or `browseVignettes("transreg")`
to open the vignette.
}
\examples{
?transreg
?predict.transreg
?coef.transreg

}
\references{
\href{https://orcid.org/0000-0001-6498-4801}{Armin Rauschenberger},
\href{https://orcid.org/0000-0002-2327-3904}{Zied Landoulsi},
\href{https://orcid.org/0000-0003-4780-8472}{Mark A. van de Wiel},
and
\href{https://orcid.org/0000-0003-3977-7469}{Enrico Glaab}
(2023).
"Penalised regression with multiple sets of prior effects".
\emph{Bioinformatics} 39(12):btad680.
\doi{10.1093/bioinformatics/btad680}.
(Click
\href{https://orbilu.uni.lu/bitstream/10993/53227/1/transreg.pdf}{here}
to access PDF.)
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rauschenberger/transreg}
  \item \url{https://rauschenberger.github.io/transreg/}
  \item Report bugs at \url{https://github.com/rauschenberger/transreg/issues}
}

}
\author{
\strong{Maintainer}: Armin Rauschenberger \email{armin.rauschenberger@uni.lu} (\href{https://orcid.org/0000-0001-6498-4801}{ORCID})

}
\keyword{documentation}
