% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgname.R
\docType{package}
\name{transreg-package}
\alias{transreg-package}
\title{Penalised regression with multiple sets of prior effects}
\description{
The R package \code{transreg} implements penalised regression
with multiple sets of prior effects.
}
\details{
Use function \code{\link[=transreg]{transreg()}} for model fitting.
Type \code{library(transreg)} and then \code{?transreg} or
\code{help("transreg")} to open its help file.

See the vignette for further examples.
Type \code{vignette("transreg")} or \code{browseVignettes("transreg")}
to open the vignette.
}
\examples{
?transreg
?predict.transreg
?coef.transreg

}
\references{
\href{https://orcid.org/0000-0001-6498-4801}{Armin Rauschenberger},
\href{https://orcid.org/0000-0002-2327-3904}{Zied Landoulsi},
\href{https://orcid.org/0000-0003-4780-8472}{Mark A. van de Wiel},
and
\href{https://orcid.org/0000-0003-3977-7469}{Enrico Glaab}
(2023).
"Penalised regression with multiple sets of prior effects".
\emph{Bioinformatics (In press).}
\doi{10.1093/bioinformatics/btad680}
\email{armin.rauschenberger@uni.lu}
}
\keyword{documentation}
