% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{transreg}
\alias{transreg}
\title{Penalised regression with multiple sets of prior coefficients}
\usage{
transreg(
  y,
  X,
  prior,
  family = "gaussian",
  alpha = 1,
  foldid = NULL,
  nfolds = 10,
  scale = "iso",
  sign = FALSE,
  switch = TRUE,
  select = TRUE
)
}
\arguments{
\item{y}{target: vector of length \eqn{n}}

\item{X}{features: matrix with \eqn{n} rows and \eqn{p} columns}

\item{prior}{prior coefficients: matrix with \eqn{p} rows and \eqn{k} columns}

\item{family}{character "gaussian", "binomial", or "poisson"}

\item{alpha}{elastic net mixing parameter (0=ridge, 1=lasso)}

\item{foldid}{fold identifiers: vector of length \eqn{n} with entries from 1 to \code{nfolds}}

\item{nfolds}{number of folds: integer}

\item{scale}{"exp" for exponential scaling (3 parameters),
"iso" for isotonic scaling (1+p parameters)}

\item{sign}{sign discovery procedure: logical}

\item{switch}{choose between positive and negative weights for each source: logical}

\item{select}{select from sources: logical}
}
\description{
Implements penalised regression with multiple sets of prior coefficients
}
\examples{
NA

}
\seealso{
Methods for objects of class \code{cornet}
include ...
}
